+++ changelog.md
@@ -15,6 +15,8 @@
 * [3534] (https://github.com/zeta-chain/node/pull/3534) - Add Sui deposit & depositAndCall
 * [3541](https://github.com/zeta-chain/node/pull/3541) - implement `MsgUpdateZRC20Name` to update the name or symbol of a ZRC20 token
 * [3520](https://github.com/zeta-chain/node/pull/3520) - SPL withdraw and call integration
+* [3439](https://github.com/zeta-chain/node/pull/3439) - use protocol contracts V2 with TON deposits
+* [3522](https://github.com/zeta-chain/node/pull/3522) - add `MsgDisableFastConfirmation` to disable fast confirmation. This message can be triggered by the emergency policy.
 * [3551](https://github.com/zeta-chain/node/pull/3551) - support for fast inbound confirmation
 
 ### Refactor